day 6
at least I didn't need to sort the map

I like the trick of using string() for the map instead of rune/byte because that's made debug prints hard in katas.

my first impulse was to make the map[]int but went with map[]bool and had to change it :(
Should have guessed part 2 would need the count.